split test cases by different types
